AnotherSpotiStatsApp/requirements.txt
appdirs==1.4.4
arrow==1.2.2
asgiref==3.2.10
astroid==2.9.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
black==20.8b1
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
Collectfast==2.2.0
cookiecutter==1.7.3
coverage==5.3
cryptography==35.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-anymail==8.5
django-autoslug==1.9.8
django-countries==7.2.1
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
django-environ==0.4.5
django-extensions==3.0.9
django-heroku==0.3.1
django-model-utils==4.0.0
django-redis==5.2.0
django-storages==1.12.3
django-stubs==1.5.0
django-test-plus==1.4.0
executing==0.8.3
factory-boy==3.0.1
Faker==13.3.2
favicon==0.7.0
filelock==3.6.0
flake8==3.8.3
gunicorn==20.1.0
identify==2.4.12
idna==3.3
iniconfig==1.1.1
ipdb==0.13.3
ipython==8.1.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.1
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==21.2.4
platformdirs==2.5.1
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.7.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
Pygments==2.11.2
pylint==2.12.2
pylint-django==2.3.0
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==6.1.0
pytest-django==3.10.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==6.0
redis==4.2.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==58.0.4
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.1
typed-ast==1.4.3
typing_extensions==4.1.1
Unidecode==1.1.1
urllib3==1.26.9
virtualenv==20.13.4
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.37.1
whitenoise==5.2.0
wrapt==1.13.3
